[DROOLS-984] @expires( -1 ) should disable Event expiration

+++ drools-core/src/main/java/org/drools/core/base/evaluators/FinishesEvaluatorDefinition.java
@@ -155,7 +155,7 @@ public class FinishesEvaluatorDefinition
         String key = isNegated + ":" + parameterText;
         FinishesEvaluator eval = this.cache.get( key );
         if ( eval == null ) {
-            Long[] params = TimeIntervalParser.parse( parameterText );
+            long[] params = TimeIntervalParser.parse( parameterText );
             eval = new FinishesEvaluator( type,
                                           isNegated,
                                           params,
@@ -214,7 +214,7 @@ public class FinishesEvaluatorDefinition
 
         public FinishesEvaluator(final ValueType type,
                                  final boolean isNegated,
-                                 final Long[] parameters,
+                                 final long[] parameters,
                                  final String paramText) {
             super( type,
                    isNegated ? FINISHES_NOT : FINISHES );
@@ -330,13 +330,13 @@ public class FinishesEvaluatorDefinition
          *
          * @param parameters
          */
-        private void setParameters(Long[] parameters) {
+        private void setParameters(long[] parameters) {
             if ( parameters == null || parameters.length == 0 ) {
                 this.endDev = 0;
             } else if ( parameters.length == 1 ) {
-                if( parameters[0].longValue() >= 0 ) {
+                if( parameters[0] >= 0 ) {
                     // defined deviation for end timestamp
-                    this.endDev = parameters[0].longValue();
+                    this.endDev = parameters[0];
                 } else {
                     throw new RuntimeException("[Finishes Evaluator]: Not possible to use negative parameter: '" + paramText + "'");
                 }
