+++ drools-core/src/main/java/org/drools/core/base/evaluators/OverlappedByEvaluatorDefinition.java
@@ -164,7 +164,7 @@ public class OverlappedByEvaluatorDefinition
         String key = isNegated + ":" + parameterText;
         OverlappedByEvaluator eval = this.cache.get( key );
         if ( eval == null ) {
-            Long[] params = TimeIntervalParser.parse( parameterText );
+            long[] params = TimeIntervalParser.parse( parameterText );
             eval = new OverlappedByEvaluator( type,
                                        isNegated,
                                        params,
@@ -223,7 +223,7 @@ public class OverlappedByEvaluatorDefinition
 
         public OverlappedByEvaluator(final ValueType type,
                               final boolean isNegated,
-                              final Long[] parameters,
+                              final long[] parameters,
                               final String paramText) {
             super( type,
                    isNegated ? NOT_OVERLAPPED_BY : OVERLAPPED_BY );
@@ -350,7 +350,7 @@ public class OverlappedByEvaluatorDefinition
          *
          * @param parameters
          */
-        private void setParameters(Long[] parameters) {
+        private void setParameters(long[] parameters) {
             if ( parameters == null || parameters.length == 0 ) {
                 // open bounded range
                 this.minDev = 1;
@@ -358,11 +358,11 @@ public class OverlappedByEvaluatorDefinition
             } else if ( parameters.length == 1 ) {
                 // open bounded ranges
                 this.minDev = 1;
-                this.maxDev = parameters[0].longValue();
+                this.maxDev = parameters[0];
             } else if ( parameters.length == 2 ) {
                 // open bounded ranges
-                this.minDev = parameters[0].longValue();
-                this.maxDev = parameters[1].longValue();
+                this.minDev = parameters[0];
+                this.maxDev = parameters[1];
             } else {
                 throw new RuntimeException( "[Overlaps Evaluator]: Not possible to use " + parameters.length + " parameters: '" + paramText + "'" );
             }
